requirements.txt
cached_property
click
dataclasses
git+https://github.com/turicas/rows.git@develop#egg=rows
lxml
openpyxl==2.5.12
python-Levenshtein
requests
requests
requests-cache
scrapy
tqdm
xlrd
xlwt
